feat: 출석 세션에 대한   api 연결
