+++ Projects/Domain/Entity/Sources/OnBoarding/SelectTeam.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import ComposableArchitecture
 
 // MARK: - 나중에
 //SelectTeam 으로 이름 변경 예정
@@ -20,5 +21,28 @@ public struct SelectTeamEntity: Equatable, Identifiable {
   ) {
     self.teamId = teamId
     self.teams = teams
+  }
+}
+
+
+public extension SelectTeamEntity {
+  var toSelectTeam: SelectTeams? {
+    switch teams {
+    case .and1: return .and1
+    case .and2: return .and2
+    case .ios1: return .ios1
+    case .ios2: return .ios2
+    case .web1: return .web1
+    case .web2: return .web2
+    case .unknown: return nil
+    }
+  }
+}
+
+extension IdentifiedArrayOf where Element == SelectTeamEntity {
+  func orderedSelectTeams() -> [SelectTeams] {
+    return self
+      .sorted { $0.teamId < $1.teamId }
+      .compactMap { $0.toSelectTeam }
   }
 }
