+++ Projects/Presentation/Management/Sources/AttendanceCheck/Reducer/AttendanceCheck.swift
@@ -43,7 +43,6 @@ public struct AttendanceCheck {
     var attendanceModel: IdentifiedArrayOf<Attendance> = .init(uniqueElements: [])
 
 
-
     public init() {
 
     }
